one more point
this is like the THIRD time a taiwanese has made me uncomfortable
well, the first two are probably just me.... but even if this were
the first time i feel uncomfortable enough to say something about it,
one time is more than you think it is.  it takes one person to ruin
the entire group!

i am not saying that i have not heard of such legalese speak though.
coming from a PERSON rather than a corporate representative, however,
this was the first time.  it's this stinging reminder that such a
person actually exists.  but if it empowers us in any way, it's that
these corporate overlords do not have any way over whatever imagination
we have of them.  like the worst thing you can do to a person -- to
assume their disposition without even bothering to ask them first --
if you do that to someone you'd usually feel remorse.  but when
someone says "We hope you understand this decision" without any
satisfactory rationale, it's not just abuse of power, and more like...
a lack of human decency and sensibility.  and it is people like those
that make me feel comfortable even ASSUMING that they did this because
they fear that if a chinese gets a foothold in their business, he would
denouce them for keeping data that would be useful to other students
and revolt one day.  they think that their neat taiwanese group was
neat and nice and don't want a chinese to "taint" their taiwanese purity.
(or prove me wrong, maybe BOTH of them thought i was an issue?  well
in that case, leaving on a good note and having my hopes up is just
a scummy practice.  i did have to forfeit some of my assignments due on
Friday just to attend their untimely meeting; the least they could do is
EXPLAIN to me whatever the f*ck was their issue.)

i'm sorry, am i really autistic?  maybe i really am.  and to be
honest... i don't really care that much about a course selection
site.  i only attend university once, and i only live my 20s once.
i couldn't care less about making data that is already accessible
but just not aggregated enough for other people.  i was thinking of
buying BUCKYMATCH.COM just to tease the owner for making that idiotic
excuse that he has not yet bought the domain when i told him to do so
WEEKS AGO (buying a domain does NOT take any time, bro. =n=) but
instead of that i went to GoDaddy --- their favorite registrar, DNS
host and email (which is basically resold from Microsoft, ironically)
--- and see if the prophecy comes true that GoDaddy claims searched-up
domains as premium.  If GoDaddy claims it as premium, well I did warn
them weeks ago!  And they decided to block me (which is like gross;
like, you must consider someone to be less than human -- like a greedy
scammer, or a child groomer -- to not trust that they would stop talking
to you even if you said so and PREVENT them from talking to you.)
so even if I wanted to warn them again I can't!  They brought this upon
themselves, and therefore only have themselves to blame. :)

truth be told, i do not think that i want to hate Taiwanese... but
at the very least THIS particular taiwanese is a stinky scumbag to me.
just for collecting student email addresses alone without considering
for a millisecond to write up a privacy policy i wish their MongoDB
gets leaked and everyone gets spammed and ends up hating them.
(I know that likely won't happen, but.... i can't wait to see their
downfall when they graduate and toss the domain in the garbage, just
like what everyone does... >:)

(gosh that was the longest commit message every -- but in the meantime,
this is much better than what Twitter lets you say!  i am glad i didn't
grow up with Twitter and social media but instead the comfort of my own
journals and text files.... :)