use more precise wording for obscure database schemas
undocumented sounds more like "nah i'm lazy, but i'd be down to
explain if you want me to!" whereas i am talking about the kind of
PROPRIETARY undocumented and is more like "nah, we're lazy.... but
even if we did write such thing, you won't be the one to see it."
or once you reverse engineer it you encounter a f**k ton of
different, competing, contradicting standards (and, for the cream
of the crop, make it INCOMPATIBLE and impossible to tell apart).
exactly like the way Adobe PSD did it decades ago!

the goal is to make it so that data so sh*tty, so obfuscated, that
the file is difficult to operate or otherwise completely inoperable
in a different system.  even IF you software must be proprietary, i
would regard it much more highly if the file format is well-documented
(although despite OOXML i am still ambivalent about people sending me
Microsoft Word files.)

oh, and OneNote exports can definitely go f**k itself. :)  this may
come back to me in not a good way if Microsoft did release their
OneNote file format spec, but just for the psychological damage of
staring at the hexdump of their garbage, knowing that my stuff
COULD be in there but cannot tell if that was the case at all
because the way i GOT the file in the first place was from F**KING
CACHE from my rooted Android phone (because guess what, Microsoft
definitely does not want you to stop using OneNote one day, and
help you export all your notes in a way that can be opened in a
different program).  it's not just about the file format; it's
their whole business practice that makes me lose hope in humanity